global_average_timeseries - N_obs data type adjusted for netCDF writing
timeseries_comparison plots - creates plots to compare timeseries across different versions of HadISDH.

utils.py - added the original QC_FLAGS array back, but commented out just in case useful in the future.

+++ EUSTACE_SST_MAT/global_average_timeseries.py
@@ -180,7 +180,7 @@ for version in ["", "_anomalies"]:
         lat_centres = ncdf_file.variables["latitude"]
         lon_centres = ncdf_file.variables["longitude"]
 
-        n_obs = utils.set_MetVar_attributes("n_obs", "Number of Observations", "Number of Observations", 1, -1, np.dtype("int16"), 0)
+        n_obs = utils.set_MetVar_attributes("n_obs", "Number of Observations", "Number of Observations", 1, -1, np.dtype("int64"), 0)
         OBS_ORDER = utils.make_MetVars(mdi, multiplier = False)
         OBS_ORDER += [n_obs]
